api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-generativeai==0.8.0
groq==0.11.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.0
